+++ src/world/SpellHandler.cpp
@@ -94,7 +94,7 @@ void WorldSession::HandleUseItemOpcode(WorldPacket& recvPacket)
     }
 
     // Let's check if the item even has that spell
-    for (int i = 0; i < 5; ++i)
+    for (uint8 i = 0; i < 5; ++i)
     {
         if (itemProto->Spells[i].Trigger == USE && itemProto->Spells[i].Id == spellId)
         {
